pyproject.toml
[tool.poetry]
name = "image-sharing-api"
version = "0.1.0"
description = ""
authors = ["Fafa Modey <uffamodey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
django = "^5.1.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
pyjwt = "^2.9.0"
pytest = "^8.3.3"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
